# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Sentinel Gateway API
  version: "0.4.3"
  description: |
    Security guardrail proxy for AI agents. Sits between users/applications and LLM
    backends, enforcing security policies on every request in real-time.

    ## Authentication

    All endpoints (except `/health` and `/health/live`) require authentication via
    one of the supported schemes.

    ## API Versioning

    - **V1** (`/v1/*`): OpenAI-compatible proxy endpoints. Stable, backward-compatible.
    - **V2** (`/v2/*`): Standalone scanning API. Date-versioned via `X-API-Version` header.

    The `X-API-Version` header is returned on all V2 responses.

    ## Rate Limiting

    Rate limits are enforced per tenant. When exceeded, `429 Too Many Requests` is returned
    with `X-RateLimit-*` headers indicating current state.

    ## Error Format

    All errors follow a consistent format:
    ```json
    {"error": {"message": "...", "type": "...", "code": "..."}}
    ```
  contact:
    name: Sentinel Gateway Team
    url: https://github.com/sentinel-gateway/sentinel-gateway
  license:
    name: GPL-3.0-or-later
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.sentinel-gateway.example.com
    description: Production
security:
  - BearerJWT: []
  - APIKey: []
...
paths:
  /v1/chat/completions:
    post:
      operationId: chatCompletions
      summary: Chat completions with guardrails
      description: |
        OpenAI-compatible chat completions endpoint with full security pipeline:
        1. Input guardrail (prompt injection, jailbreak detection)
        2. IOC check (malicious URLs/IPs/hashes)
        3. Forward to backend LLM
        4. Tool policy enforcement (RBAC on tool calls)
        5. Output filter (secret/PII redaction)

        Supports both streaming (SSE) and non-streaming responses.
      tags: [Proxy]
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/AgentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
            examples:
              basic:
                summary: Basic chat request
                value:
                  model: "gpt-4"
                  messages:
                    - role: "user"
                      content: "Explain quantum computing in simple terms."
              with_tools:
                summary: Request with tool definitions
                value:
                  model: "gpt-4"
                  messages:
                    - role: "user"
                      content: "Search for the latest news about AI safety."
                  tools:
                    - type: function
                      function:
                        name: web_search
                        parameters:
                          type: object
                          properties:
                            query:
                              type: string
              streaming:
                summary: Streaming request
                value:
                  model: "gpt-4"
                  messages:
                    - role: "user"
                      content: "Write a short poem."
                  stream: true
      responses:
        "200":
          description: Successful completion
          headers:
            X-Request-ID:
              description: Unique request identifier for correlation
              schema:
                type: string
                format: uuid
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
            X-Quota-Remaining-Tokens:
              $ref: "#/components/headers/X-Quota-Remaining-Tokens"
            X-Quota-Reset:
              $ref: "#/components/headers/X-Quota-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
              examples:
                success:
                  summary: Successful response
                  value:
                    id: "chatcmpl-abc123"
                    object: "chat.completion"
                    created: 1719001234
                    model: "gpt-4"
                    choices:
                      - index: 0
                        message:
                          role: "assistant"
                          content: "Quantum computing uses quantum bits..."
                        finish_reason: "stop"
                    usage:
                      prompt_tokens: 12
                      completion_tokens: 45
                      total_tokens: 57
            text/event-stream:
              schema:
                type: string
                description: Server-sent events stream (SSE)
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_json:
                  value:
                    error:
                      message: "Invalid JSON body"
                      type: "validation_error"
                      code: "invalid_request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Content blocked by guardrail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                blocked:
                  summary: Prompt injection detected
                  value:
                    error:
                      message: "Request blocked: security policy violation"
                      type: "security_violation"
                      code: "input_guardrail"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                too_large:
                  value:
                    error:
                      message: "Request body too large (max 10MB)"
                      type: "validation_error"
                      code: "body_too_large"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          description: Backend LLM unavailable or returned an error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                backend_error:
                  value:
                    error:
                      message: "Backend unavailable"
                      type: "backend_error"
                      code: "upstream_error"
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with required claims: `sub` (subject), `aud` (must include
        "sentinel-proxy"). Optional claims: `role` (for admin endpoints),
        `tenant_id`, `agent_id`.
    APIKey:
      type: http
      scheme: bearer
      description: |
        API key passed as Bearer token. Must be listed in `SENTINEL_API_KEYS`
        configuration. Tenant and agent IDs are extracted from request headers.
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_\\-]{1,64}$"
        example: "default-corp"
      description: Tenant identifier for multi-tenant routing and policy resolution
    AgentID:
      name: X-Agent-ID
      in: header
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_\\-]{1,64}$"
        example: "support-bot"
      description: Agent identifier for backend routing and tool policy resolution
...
  headers:
    X-RateLimit-Limit:
      description: Maximum requests allowed per window
      schema:
        type: integer
        example: 60
    X-RateLimit-Remaining:
      description: Remaining requests in current window
      schema:
        type: integer
        example: 42
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        format: int64
        example: 1719001260
    X-Quota-Remaining-Tokens:
      description: Remaining daily token quota for the tenant
      schema:
        type: integer
        example: 950000
    X-Quota-Reset:
      description: Unix timestamp when the daily quota resets (midnight UTC)
      schema:
        type: integer
        format: int64
        example: 1719100800
...
  responses:
    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing_auth:
              summary: No credentials provided
              value:
                error:
                  message: "Missing or invalid Authorization header"
                  type: "authentication_error"
                  code: "missing_auth"
            invalid_token:
              summary: Expired or malformed JWT
              value:
                error:
                  message: "Invalid or expired token"
                  type: "authentication_error"
                  code: "invalid_token"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
        Retry-After:
          description: Seconds until the rate limit resets
          schema:
            type: integer
            example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            rate_limited:
              value:
                error:
                  message: "Rate limit exceeded (60 requests/minute)"
                  type: "rate_limit_error"
                  code: "rate_limited"
  schemas:
...
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: Model identifier (resolved by agent registry)
          example: "gpt-4"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
          minItems: 1
          description: Conversation messages
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ToolDefinition"
          description: Available tool definitions
        tool_choice:
          oneOf:
            - type: string
              enum: ["none", "auto", "required"]
            - type: object
              properties:
                type:
                  type: string
                  enum: ["function"]
                function:
                  type: object
                  properties:
                    name:
                      type: string
          description: Tool call control
        temperature:
          type: number
          minimum: 0
          maximum: 2
          description: Sampling temperature
        max_tokens:
          type: integer
          minimum: 1
          description: Maximum tokens to generate
        stream:
          type: boolean
          default: false
          description: Enable SSE streaming
    Message:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: ["system", "user", "assistant", "tool"]
          description: Message role
        content:
          type: string
          nullable: true
          description: Message content (null for tool_calls-only assistant messages)
        tool_calls:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
                enum: ["function"]
              function:
                type: object
                properties:
                  name:
                    type: string
                  arguments:
                    type: string
          description: Tool calls (assistant messages only)
        tool_call_id:
          type: string
          description: ID of the tool call this message responds to (tool role only)
    ToolDefinition:
      type: object
      properties:
        type:
          type: string
          enum: ["function"]
        function:
          type: object
          properties:
            name:
              type: string
            description:
              type: string
            parameters:
              type: object
          required: [name]
...
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique completion identifier
        object:
          type: string
          enum: ["chat.completion"]
        created:
          type: integer
          format: int64
          description: Unix timestamp of creation
        model:
          type: string
          description: Model used for completion
        choices:
          type: array
          items:
            $ref: "#/components/schemas/Choice"
        usage:
          $ref: "#/components/schemas/Usage"
    Choice:
      type: object
      properties:
        index:
          type: integer
        message:
          $ref: "#/components/schemas/Message"
        finish_reason:
          type: string
          enum: ["stop", "length", "tool_calls", "content_filter"]
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: object
              required: [message, type, code]
              properties:
                message:
                  type: string
                  description: Human-readable error description
                type:
                  type: string
                  description: Error category
                  enum:
                    - validation_error
                    - authentication_error
                    - security_violation
                    - rate_limit_error
                    - quota_exceeded
                    - backend_error
                code:
                  type: string
                  description: Machine-readable error code
            - type: string
              description: Simple error message (legacy format)